rename gits package to meitrex
