+++ Dockerfile
@@ -21,4 +21,4 @@ ARG DEPENDENCY=/workspace/app/build/dependency
 COPY --from=build ${DEPENDENCY}/BOOT-INF/lib /app/lib
 COPY --from=build ${DEPENDENCY}/META-INF /app/META-INF
 COPY --from=build ${DEPENDENCY}/BOOT-INF/classes /app
-ENTRYPOINT ["java","-cp","app:app/lib/*","de.unistuttgart.iste.gits.user_service.UserServiceApplication"]
+ENTRYPOINT ["java","-cp","app:app/lib/*","de.unistuttgart.iste.meitrex.user_service.UserServiceApplication"]
