+++ src/main/java/de/unistuttgart/iste/meitrex/user_service/config/KeycloakConfiguration.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.iste.gits.user_service.config;
+package de.unistuttgart.iste.meitrex.user_service.config;
 
 import lombok.RequiredArgsConstructor;
 import org.keycloak.admin.client.Keycloak;
