+++ src/main/java/de/unistuttgart/iste/meitrex/user_service/controller/UserController.java
@@ -1,9 +1,9 @@
-package de.unistuttgart.iste.gits.user_service.controller;
+package de.unistuttgart.iste.meitrex.user_service.controller;
 
-import de.unistuttgart.iste.gits.generated.dto.PublicUserInfo;
-import de.unistuttgart.iste.gits.generated.dto.UserInfo;
-import de.unistuttgart.iste.gits.user_service.service.UserService;
 import de.unistuttgart.iste.meitrex.common.user_handling.LoggedInUser;
+import de.unistuttgart.iste.meitrex.generated.dto.PublicUserInfo;
+import de.unistuttgart.iste.meitrex.generated.dto.UserInfo;
+import de.unistuttgart.iste.meitrex.user_service.service.UserService;
 import graphql.schema.DataFetchingEnvironment;
 import lombok.RequiredArgsConstructor;
 import org.springframework.graphql.data.method.annotation.*;
