+++ src/main/java/de/unistuttgart/iste/meitrex/user_service/mapper/RealmMapper.java
@@ -1,7 +1,7 @@
-package de.unistuttgart.iste.gits.user_service.mapper;
+package de.unistuttgart.iste.meitrex.user_service.mapper;
 
-import de.unistuttgart.iste.gits.generated.dto.GlobalUserRole;
 import de.unistuttgart.iste.meitrex.common.user_handling.LoggedInUser;
+import de.unistuttgart.iste.meitrex.generated.dto.GlobalUserRole;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
 
@@ -12,16 +12,16 @@ public class RealmMapper {
 
     private final ModelMapper modelMapper;
 
-    public RealmMapper(ModelMapper modelMapper){
+    public RealmMapper(ModelMapper modelMapper) {
         this.modelMapper = modelMapper;
     }
 
-    public List<GlobalUserRole> internalRolesToGraphQlRoles (Set<LoggedInUser.RealmRole> realmRoleSet){
+    public List<GlobalUserRole> internalRolesToGraphQlRoles(Set<LoggedInUser.RealmRole> realmRoleSet) {
         return realmRoleSet.stream().map(role -> modelMapper.map(role, GlobalUserRole.class)).toList();
     }
 
-    public List<GlobalUserRole> keycloakRolesToGraphQlRoles(List<String> keycloakRoles){
-        if (keycloakRoles == null){
+    public List<GlobalUserRole> keycloakRolesToGraphQlRoles(List<String> keycloakRoles) {
+        if (keycloakRoles == null) {
             return Collections.emptyList();
         }
